movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
numpy==1.19.5
pandas==1.2.0
factory-boy==3.2.0
django-extensions==3.1.0
django-print-sql==2018.3.6